Refactoring + adding p_grid cmd-argument.
