+++ examples/conflux_miniapp.cpp
@@ -19,7 +19,10 @@ int main(int argc, char *argv[]) {
         ("b,block_size",
             "block size",
             cxxopts::value<int>()->default_value("256"))
-        ("p,print_limit",
+        ("p, p_grid",
+            "3D-process decomposition.", 
+            cxxopts::value<std::vector<int>>()->default_value("-1,-1,-1"))
+        ("l,print_limit",
             "limit for printing the final result.", 
             cxxopts::value<int>()->default_value("20"))
         ("r,n_rep",
@@ -41,28 +44,48 @@ int main(int argc, char *argv[]) {
     auto b = result["block_size"].as<int>();
     auto n_rep = result["n_rep"].as<int>();
     auto print_limit = result["print_limit"].as<int>();
+    auto p_grid = result["p_grid"].as<std::vector<int>>();
 
     MPI_Init(&argc, &argv);
-    int rank, size;
-    MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-    MPI_Comm_size(MPI_COMM_WORLD, &size);
+    int rank, P;
+    MPI_Comm comm = MPI_COMM_WORLD;
+    MPI_Comm_rank(comm, &rank);
+    MPI_Comm_size(comm, &P);
+    conflux::lu_params<double> params;
 
-    auto gv = conflux::GlobalVars<dtype>(M, N, b, size);
+    int newP = P;
 
-    std::cout << "Rank: " << rank << ", M: " << gv.M << ", N: " << gv.N << ", P:" << gv.P
-              << ", v:" << gv.v << ", Px:" << gv.Px << ", Py: " << gv.Py << ", Pz: " << gv.Pz
-              << ", Nt: " << gv.Nt
-              << ", tA11x: " << gv.tA11x << ", tA11y: " << gv.tA11y << std::endl;
+    if (p_grid[0] <= 0 || p_grid[1] <= 0 || p_grid[2] <= 0) {
+        params = conflux::lu_params<double>(M, N, b, P);
+    } else {
+        params = conflux::lu_params<double>(M, N, b, p_grid[0], p_grid[1], p_grid[2]);
+        newP = p_grid[0] * p_grid[1] * p_grid[2];
+    }
 
-    std::vector<double> C(gv.M * gv.N);
-    std::vector<double> Perm(gv.M * gv.M);
+    if (newP < P) {
+        std::vector<int> ranks;
+        for (int i = 0; i < newP; ++i) {
+            ranks.push_back(i);
+        }
+        comm = conflux::create_comm(comm, ranks);
+    }
+
+    if (rank == 0) {
+        std::cout << "Rank: " << rank << ", M: " << params.M << ", N: " << params.N << ", P:" << params.P
+                  << ", v:" << params.v << ", Px:" << params.Px << ", Py: " << params.Py << ", Pz: " << params.Pz
+                  << ", Nt: " << params.Nt
+                  << ", tA11x: " << params.tA11x << ", tA11y: " << params.tA11y << std::endl;
+    }
+
+    std::vector<double> C(params.M * params.N);
+    std::vector<double> Perm(params.M * params.M);
 
     for (int i = 0; i < n_rep; ++i) {
         PC();
-        conflux::LU_rep<dtype>(gv.matrix.data(), 
+        conflux::LU_rep<double>(params.matrix.data(), 
                                C.data(), 
                                Perm.data(), 
-                               gv, 
+                               params, 
                                MPI_COMM_WORLD);  
         // print the profiler data
         if (rank == 0) {
@@ -71,8 +94,8 @@ int main(int argc, char *argv[]) {
 
 #ifdef CONFLUX_WITH_VALIDATION
         if (rank == 0) {
-            auto M = gv.M;
-            auto N = gv.N;
+            auto M = params.M;
+            auto N = params.N;
             std::vector<dtype> L(N*N);
             std::vector<dtype> U(N*N);
             for (auto i = 0; i < N; ++i) {
@@ -86,7 +109,7 @@ int main(int argc, char *argv[]) {
             }
 
             // mm<dtype>(L, U, C, N, N, N);
-            // gemm<dtype>(PP, gv.matrix, C, -1.0, 1.0, N, N, N);
+            // gemm<dtype>(PP, params.matrix, C, -1.0, 1.0, N, N, N);
             cblas_dgemm(CblasRowMajor, CblasNoTrans, CblasNoTrans, N, N, N,
                         1.0, &L[0], N, &U[0], N, 0.0, C.data(), N);
 
@@ -102,7 +125,7 @@ int main(int argc, char *argv[]) {
             }
 
             cblas_dgemm(CblasRowMajor, CblasNoTrans, CblasNoTrans, N, N, N,
-                        -1.0, Perm.data(), N, gv.matrix.data(), N, 1.0, C.data(), N);
+                        -1.0, Perm.data(), N, params.matrix.data(), N, 1.0, C.data(), N);
             if (rank == 0 && std::max(M, N) < print_limit){ 
                 std::cout << "\nL*U - P*A:\n";
                 conflux::print_matrix(C.data(), 0, M, 0, N, N);
@@ -119,6 +142,10 @@ int main(int argc, char *argv[]) {
 #endif
     }
 
+    // if not MPI_COMM_WORLD, deallocate it
+    if (newP < P) {
+        MPI_Comm_free(&comm);
+    }
     MPI_Finalize();
 
     return 0;
